+++ kite-tools/src/test/java/org/kitesdk/cli/commands/TestCSVImportCommandCluster.java
@@ -40,7 +40,7 @@ import org.kitesdk.data.Dataset;
 import org.kitesdk.data.Datasets;
 import org.kitesdk.data.MiniDFSTest;
 import org.kitesdk.data.TestHelpers;
-import org.kitesdk.data.spi.URIBuilder;
+import org.kitesdk.data.URIBuilder;
 import org.kitesdk.data.spi.filesystem.DatasetTestUtilities;
 import org.slf4j.Logger;
 
@@ -93,7 +93,7 @@ public class TestCSVImportCommandCluster extends MiniDFSTest {
   public void setup() throws Exception {
     TestUtil.run("create", datasetName, "-r", repoURI, "-s", avsc);
 
-    this.dataset = Datasets.load(new URIBuilder(repoURI, "default", datasetName).build(),
+    this.dataset = Datasets.load(URIBuilder.build(repoURI, "default", datasetName),
         GenericData.Record.class);
     this.console = mock(Logger.class);
     this.command = new CSVImportCommand(console);
